+++ src/main/java/io/github/hapjava/characteristics/impl/common/ActiveIdentifierCharacteristic.java
@@ -1,0 +1,29 @@
+package io.github.hapjava.characteristics.impl.common;
+
+import io.github.hapjava.characteristics.ExceptionalConsumer;
+import io.github.hapjava.characteristics.HomekitCharacteristicChangeCallback;
+import io.github.hapjava.characteristics.impl.base.IntegerCharacteristic;
+import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
+import java.util.function.Consumer;
+import java.util.function.Supplier;
+
+public class ActiveIdentifierCharacteristic extends IntegerCharacteristic {
+
+  public ActiveIdentifierCharacteristic(
+      Supplier<CompletableFuture<Integer>> getter,
+      ExceptionalConsumer<Integer> setter,
+      Consumer<HomekitCharacteristicChangeCallback> subscriber,
+      Runnable unsubscriber) {
+    super(
+        "000000E7-0000-1000-8000-0026BB765291",
+        "active identifier",
+        0,
+        1000,
+        "",
+        Optional.of(getter),
+        Optional.of(setter),
+        Optional.of(subscriber),
+        Optional.of(unsubscriber));
+  }
+}